Treat Sd as pelagic for encounters
I added the is_small property to fish (similar to is_demersal), and added
is_small=False to both zooplankton and benthic_prey so that logic on whether to
use 1-t_pelagic can depend on size of the demersal prey.

Also updated output from Matlab (changed a few preference rates to match what
was in the python code).