Remove assumption of a single zooplankton
compute_tracer_tendency() takes an additional argument named first_fish_id,
which is the index of the first fish in the group dimension. When looping over
enumerate(member_obj_link), i is the index of the object in the group dimension
so i - first_fish_id is the index into the fish dimension.