Add pytest for cyclic interpolation
New test initializes two instances of the idealized test case, one starting on
0001-01-01 and one starting on 0002-01-01, runs both for a single time step,
and expects biomass to be identical between the two. Currently this test fails,
because the forcing is only defined for year 1 and using xr.DataArray.interp()
in year 2 returns a nan which then propogates throughout the biomass array.

Proper implementation of an option for cyclic interpolation should allow this
test to pass.