Move cyclic forcing from run() to __init__()
config_testcase() will determine whether to set cyclic_forcing=True based on
which forcing profile is being used; config_from_netcdf() takes an argument
(default is False).

Also added two new tests that run a single day, and needed to shoehorn in a way
to make the forcing actually be cyclic so this commit contains some kludgy code
that will be cleaned up in coming commits.