First pass at API mimicking CESM streams
We want to eventually point FEISTY at a set of stream files to use for forcing,
and this commit introduces that API. The feisty-config.*.yml files in examples/
show the type of dictionary the config_and_run_from_yaml() interface expects,
with the forcing key dictating where to read the forcing from.

I updated FOSI_cesm.ipynb and Multiple Forcing Files.ipynb in examples/ and
also updated the notebooks in docs/ because I needed to make some small changes
to other APIs (e.g. instead of passing in nsteps, we pass in start_date and
end_date).