Moved create_restart_file() to data_wrangling.py
Instead of being a method on the _offline_driver class, it makes more sense to
have a free-standing function that takes a dataset as an argument (because the
_offline_driver object is used inside of map_blocks(), so it shouldn't be
responsible for writing the restart file).

I still need to make sure there is a way to initialize the model from the
restart file.