More API cleanup to support highres runs (#45)
* Big changes to support highres runs

The biggest change is that the notebooks now keep the data on the dask cluster,
rather than copying everything back to the main task. Also some small
organizational changes (adding highres stream files, using daily forcing
instead of monthly for the companion run, etc)

* Moved create_restart_file() to data_wrangling.py

Instead of being a method on the _offline_driver class, it makes more sense to
have a free-standing function that takes a dataset as an argument (because the
_offline_driver object is used inside of map_blocks(), so it shouldn't be
responsible for writing the restart file).

I still need to make sure there is a way to initialize the model from the
restart file.

* Adding default channel to conda env file

I don't know if this will help, but Github Actions are failing when conda tries
to create the environment, and it might be because some packages are on the
default channel instead of conda-forge?

* Updating channels in CI yaml files

Still trying to track down why github actions are failing

* Fix formatting in YAML

* Updating action/cache to v3

* More YAML tinkering

I had the channel listed as default rather than defaults - I don't think this
will fix the issue, but consistency is good

* Maybe updating conda will fix CI?

* Stop using cache for conda in CI

Also modified how github actions installs mamba

* Skip mamba in CI testing

* Try micromamba instead of setup-miniconda

Still tinkering with the Github Actions setup...

* Read IC from file using nlat, nlon dimensions

generate_ic_ds_for_feisty() was not set up to read a netcdf file that had
(nlat, nlon) dimensions instead of just an X dimension. With this change, the
following two scenarios are bit-for-bit:

1. Initializing with default constant values and running for two years
2. Running for one year, calling create_restart_file(), initializing from the
   restart file and running for one more year

* New notebook to run companion 1 year at a time

Needed to update how we write restarts, ended up refactoring so that writing
restarts and writing history files use the same logic to determine whether to
call to_netcdf() or to_zarr().

* Update highres forcing to use latest output

Forcing is now available from 1980 - 2021, so the example notebook uses 1980
instead of 1964.

* New notebook to run highres 1 year at a time

This one only runs for 2 years (and I modified the companion notebook to run
for 5 years instead of 6)

* Add streams for spinning up initial conditions

These need to use forcing.use_cyclic_forcing=True in the config file, but I
think there's something wrong with config_and_run_from_yaml because that
setting does not seem to be setting for forcing correctly

* Update highres chunk sizes and worker count

Increased num_workers from 48 to 72, and also changed chunksize from 100x100 to
8x1200; multi-year runs are now running ~15 min / year instead of 20.

* Move history and restart files to scratch

Just done in the companion run for now as proof-of-concept. I want to move the
I/O calls into config_and_run_from_yaml() based on what the companion notebook
is doing

* File output handled by config_and_run_from_yaml()

Cleaned up the YAML schema and reran the Multiple Year / Forced From Timeseries
notebooks.

* Several forcing improvements

1. Add flag to config YAML to convert forcing from POP units to FEISTY units
2. Add option to stream YAML to allow offset in calendar (using offset of 1 day
   applies POP forcing from YYYY-01-02 0:00:00, which is end of time interval,
   to Jan 1st)

Also added spun-up initial conditions for multi-year runs and removed large
pelagics grazing on medium demersals from food web.

* Update linting github action

bump versions of checkout, setup-python, and pre-commit/action all to 3

* Updating isort version in pre-commit config

I think an outdated version is responsible for Github Action failure

* Rerun Forced Timeseries notebooks

Also cleaned up markdown to be more accurate

* FOSI_CESM.ipynb works again

Fixed the configuration and cleaned up the notebook itself.
Also removed unnecessary lines from the other two feisty-config files.

* rerun docs/driver-example.ipynb

* Update offset for forcing time

Two big changes:

1. Changed the sign convention, so we use -1 instead of 1 to signify that the
forcing time should be a day earlier than the provided date
2. Changed the offset from -1 to -0.5 to signify the true date for the POP
forcing is off by 12 hours rather than a full day

I only reran a single notebook; we will want to rerun the full spinup but I
think we'll need to do that after addressing the issue of uncontrolled growth /
loss of biomass